Conflicts:
	ImageIllustrator/src/Modele/Modele.java
	ImageIllustrator/src/Vue/InterfaceGraphique.java

+++ ImageIllustrator/src/Modele/FiltreConvolution.java
@@ -2,15 +2,15 @@ package Modele;
 
 public class FiltreConvolution {
 	
-	public static int[][] createFiltreMoyenne(int taille)
+	public static float[][] createFiltreMoyenne(int taille)
 	{
-		int[][] filtre = new int[taille][taille];
+		float[][] filtre = new float[taille][taille];
 		
 		for(int i = 0; i<taille; i++)
 		{
 			for(int j=0; j<taille; j++)
 			{
-				filtre[i][j]=1;
+				filtre[i][j]=(float)1.0;
 			}
 		}
 		return filtre;
